requirements/base.txt
Django==3.2.11
django-environ==0.8.1
psycopg2==2.9.3
djangorestframework==3.13.1

celery==5.2.3
django-celery-results==2.2.0
django-celery-beat==2.2.1
redis

whitenoise==5.3.0

django-filter==21.1
django-cors-headers==3.11.0
django-extensions==3.1.5

drf-jwt==1.19.2
djangorestframework-simplejwt
django-smtp-ssl
django-after-response
boto3
django-storages

django-parler==2.2
django-rosetta==0.9.6
